Delta Kernel prototype
This tool generates a visual report detailing changes in both
header files and source code between two Linux kernel versions
(tags in the Linux kernel repository: old_tag and new_tag).

The tool's innovation is that it performs the diff on a
configured `old_tag` to help developers understand the specific
comparison between versions.

The gaps as of this commit.
- If kconfig logic changes between tags
- Object code is not verified to match
- One subsystem per analysis
  (needs git blame optimization to expand to more)

See README.md for more details.

Signed-off-by: Yizhen Zhang <zhangyizhenyizen@gmail.com>
Signed-off-by: Matthew Weber <matthew.l.weber3@boeing.com>